333: moved two classes to config package as they are config-specific
