+++ extensions/servlet/src/main/java/com/stormpath/sdk/servlet/config/RegisterEnabledPredicate.java
@@ -1,4 +1,4 @@
-package com.stormpath.sdk.servlet.application;
+package com.stormpath.sdk.servlet.config;
 
 import com.stormpath.sdk.application.Application;
 import com.stormpath.sdk.lang.BiPredicate;
