+++ extensions/servlet/src/main/java/com/stormpath/sdk/servlet/mvc/LoginController.java
@@ -21,8 +21,8 @@ import com.stormpath.sdk.lang.Assert;
 import com.stormpath.sdk.oauth.AccessTokenResult;
 import com.stormpath.sdk.servlet.application.ApplicationResolver;
 import com.stormpath.sdk.servlet.config.Config;
-import com.stormpath.sdk.servlet.application.RegisterEnabledResolver;
-import com.stormpath.sdk.servlet.application.RegisterEnabledPredicate;
+import com.stormpath.sdk.servlet.config.RegisterEnabledResolver;
+import com.stormpath.sdk.servlet.config.RegisterEnabledPredicate;
 import com.stormpath.sdk.servlet.form.Form;
 import com.stormpath.sdk.servlet.http.Resolver;
 import com.stormpath.sdk.servlet.http.Saver;
