+++ extensions/servlet/src/main/resources/com/stormpath/sdk/servlet/config/web.stormpath.properties
@@ -353,4 +353,4 @@ stormpath.web.login.preHandler=com.stormpath.sdk.servlet.mvc.DisabledWebHandler
 stormpath.web.login.postHandler=com.stormpath.sdk.servlet.mvc.DisabledWebHandler
 stormpath.web.register.preHandler=com.stormpath.sdk.servlet.mvc.DisabledWebHandler
 stormpath.web.register.postHandler=com.stormpath.sdk.servlet.mvc.DisabledWebHandler
-stormpath.web.register.enabled.predicate=com.stormpath.sdk.servlet.application.RegisterEnabledPredicate
+stormpath.web.register.enabled.predicate=com.stormpath.sdk.servlet.config.RegisterEnabledPredicate
