+++ extensions/servlet/src/test/groovy/com/stormpath/sdk/servlet/config/RegisterEnabledPredicateTest.groovy
@@ -1,4 +1,4 @@
-package com.stormpath.sdk.servlet.application
+package com.stormpath.sdk.servlet.config
 
 import com.stormpath.sdk.application.Application
 import org.testng.annotations.Test
