+++ extensions/servlet/src/test/groovy/com/stormpath/sdk/servlet/config/RegisterEnabledResolverTest.groovy
@@ -1,7 +1,8 @@
-package com.stormpath.sdk.servlet.application
+package com.stormpath.sdk.servlet.config
 
 import com.stormpath.sdk.application.Application
 import com.stormpath.sdk.directory.AccountStore
+import com.stormpath.sdk.servlet.application.ApplicationResolver
 import org.testng.annotations.Test
 
 import javax.servlet.http.HttpServletRequest
